+++ argopy/data_fetchers/gdac_data.py
@@ -15,12 +15,11 @@ import logging
 from typing import Literal
 
 from ..utils.format import argo_split_path
-from ..utils.decorators import deprecated
 from ..options import OPTIONS, check_gdac_option, PARALLEL_SETUP
 from ..errors import DataNotFound
 from ..stores import ArgoIndex, has_distributed, distributed
 from .proto import ArgoDataFetcherProto
-from .gdac_data_processors import pre_process_multiprof, filter_points
+from .gdac_data_processors import pre_process_multiprof
 
 
 log = logging.getLogger("argopy.gdac.data")
@@ -62,7 +61,7 @@ class GDACArgoDataFetcher(ArgoDataFetcherProto):
         cachedir: str = "",
         parallel: bool = False,
         progress: bool = False,
-        dimension: Literal['point', 'profile'] = "point",
+        dimension: Literal["point", "profile"] = "point",
         errors: str = "raise",
         api_timeout: int = 0,
         **kwargs
@@ -244,75 +243,6 @@ class GDACArgoDataFetcher(ArgoDataFetcherProto):
         self.fs.clear_cache()
         return self
 
-    @deprecated(
-        "Not serializable, please use 'gdac_data_processors.pre_process_multiprof'",
-        version="1.0.0",
-    )
-    def _preprocess_multiprof(self, ds):
-        """Pre-process one Argo multi-profile file as a collection of points
-
-        Parameters
-        ----------
-        ds: :class:`xarray.Dataset`
-            Dataset to process
-
-        Returns
-        -------
-        :class:`xarray.Dataset`
-
-        """
-        # Remove raw netcdf file attributes and replace them with argopy ones:
-        raw_attrs = ds.attrs
-        ds.attrs = {}
-        ds.attrs.update({"raw_attrs": raw_attrs})
-
-        # Rename JULD and JULD_QC to TIME and TIME_QC
-        ds = ds.rename(
-            {"JULD": "TIME", "JULD_QC": "TIME_QC", "JULD_LOCATION": "TIME_LOCATION"}
-        )
-        ds["TIME"].attrs = {
-            "long_name": "Datetime (UTC) of the station",
-            "standard_name": "time",
-        }
-
-        # Cast data types:
-        ds = ds.argo.cast_types()
-
-        # Enforce real pressure resolution : 0.1 db
-        for vname in ds.data_vars:
-            if "PRES" in vname and "QC" not in vname:
-                ds[vname].values = np.round(ds[vname].values, 1)
-
-        # Remove variables without dimensions:
-        # todo: We should be able to find a way to keep them somewhere in the data structure
-        for v in ds.data_vars:
-            if len(list(ds[v].dims)) == 0:
-                ds = ds.drop_vars(v)
-
-        ds = (
-            ds.argo.profile2point()
-        )  # Default output is a collection of points, along N_POINTS
-
-        if self.dataset_id == "phy":
-            ds.attrs["DATA_ID"] = "ARGO"
-        if self.dataset_id == "bgc":
-            ds.attrs["DATA_ID"] = "ARGO-BGC"
-        ds.attrs["DOI"] = "http://doi.org/10.17882/42182"
-        ds.attrs["Fetched_from"] = self.server
-        try:
-            ds.attrs["Fetched_by"] = getpass.getuser()
-        except:  # noqa: E722
-            ds.attrs["Fetched_by"] = "anonymous"
-        ds.attrs["Fetched_date"] = pd.to_datetime("now", utc=True).strftime("%Y/%m/%d")
-        ds.attrs["Fetched_constraints"] = self.cname()
-        ds.attrs["Fetched_uri"] = ds.encoding["source"]
-        ds = ds[np.sort(ds.data_vars)]
-
-        if self._post_filter_points:
-            ds = self.filter_points(ds)
-
-        return ds
-
     def pre_process(self, ds, *args, **kwargs):
         return pre_process_multiprof(ds, *args, **kwargs)
 
@@ -321,7 +251,7 @@ class GDACArgoDataFetcher(ArgoDataFetcherProto):
         errors: str = "ignore",
         concat: bool = True,
         concat_method: Literal["drop", "fill"] = "fill",
-        dimension: Literal['point', 'profile'] = "",
+        dimension: Literal["point", "profile"] = "",
     ):
         """Load Argo data and return a :class:`xarray.Dataset`
 
@@ -405,7 +335,9 @@ class GDACArgoDataFetcher(ArgoDataFetcherProto):
             results = results.argo.cast_types(overwrite=False)
 
             # Meta-data processing for a single merged dataset:
-            results = results.assign_coords({'N_POINTS': np.arange(0, len(results['N_POINTS']))})
+            results = results.assign_coords(
+                {"N_POINTS": np.arange(0, len(results["N_POINTS"]))}
+            )
             results = results.sortby("TIME")
 
             # Remove netcdf file attributes and replace them with simplified argopy ones:
@@ -414,7 +346,9 @@ class GDACArgoDataFetcher(ArgoDataFetcherProto):
 
                 results.attrs = {}
                 if "Processing_history" in raw_attrs:
-                    results.attrs.update({"Processing_history": raw_attrs["Processing_history"]})
+                    results.attrs.update(
+                        {"Processing_history": raw_attrs["Processing_history"]}
+                    )
                     raw_attrs.pop("Processing_history")
                 results.argo.add_history("URI merged with '%s'" % concat_method)
 
@@ -445,22 +379,6 @@ class GDACArgoDataFetcher(ArgoDataFetcherProto):
             for ds in results:
                 ds.attrs = dict(sorted(ds.attrs.items()))
         return results
-
-    @deprecated(
-        "Not serializable, please use 'gdac_data_processors.filter_points'",
-        version="1.0.0",
-    )
-    def filter_points(self, ds):
-        if hasattr(self, "BOX"):
-            access_point = "BOX"
-            access_point_opts = {"BOX": self.BOX}
-        elif hasattr(self, "CYC"):
-            access_point = "CYC"
-            access_point_opts = {"CYC": self.CYC}
-        elif hasattr(self, "WMO"):
-            access_point = "WMO"
-            access_point_opts = {"WMO": self.WMO}
-        return filter_points(ds, access_point=access_point, **access_point_opts)
 
     def transform_data_mode(self, ds: xr.Dataset, **kwargs):
         """Apply xarray argo accessor transform_data_mode method"""
@@ -541,10 +459,10 @@ class Fetch_wmo(GDACArgoDataFetcher):
         # Get list of files to load:
         if not hasattr(self, "_list_of_argo_files"):
             if self.CYC is None:
-                URIs = self.indexfs.search_wmo(self.WMO, nrows=self._nrows).uri
+                URIs = self.indexfs.query.wmo(self.WMO, nrows=self._nrows).uri
                 self._list_of_argo_files = self.uri_mono2multi(URIs)
             else:
-                self._list_of_argo_files = self.indexfs.search_wmo_cyc(
+                self._list_of_argo_files = self.indexfs.query.wmo_cyc(
                     self.WMO, self.CYC, nrows=self._nrows
                 ).uri
 
@@ -598,11 +516,9 @@ class Fetch_box(GDACArgoDataFetcher):
         # Get list of files to load:
         if not hasattr(self, "_list_of_argo_files"):
             if len(self.indexBOX) == 4:
-                URIs = self.indexfs.search_lat_lon(self.indexBOX, nrows=self._nrows).uri
+                URIs = self.indexfs.query.lon_lat(self.indexBOX, nrows=self._nrows).uri
             else:
-                URIs = self.indexfs.search_lat_lon_tim(
-                    self.indexBOX, nrows=self._nrows
-                ).uri
+                URIs = self.indexfs.query.box(self.indexBOX, nrows=self._nrows).uri
 
             if len(URIs) > 25:
                 self._list_of_argo_files = self.uri_mono2multi(URIs)