+++ argopy/stores/float/argo_float.py
@@ -2,7 +2,7 @@
 If client is online (connected to the web) we work with the 'online' implementation
 otherwise we fall back on an offline implementation.
 
-The choice is really meaningfull when the client is using a local host. In this case
+The choice is really meaningful when the client is using a local host. In this case
 we don't know if client intends to be online or offline, so we check and implement.
 
 """
@@ -43,14 +43,16 @@ class ArgoFloat(FloatStore):
         af = ArgoFloat(WMO, host='s3')     # Shortcut for s3://argo-gdac-sandbox/pub
 
     .. code-block:: python
-        :caption: Load/read GDAC netcdf files as a :class:`xarray.Dataset`
+        :caption: Load/read GDAC netcdf files as a :class:`xarray.Dataset` or :class:`netCDF4.Dataset`
 
-        af.list_dataset() # Return a dictionary with all available datasets for this float
+        af.ls_dataset() # Return a dictionary with all available datasets for this float
         ds = af.open_dataset('prof') # Use keys from the available datasets dictionary
         ds = af.open_dataset('meta')
         ds = af.open_dataset('tech')
         ds = af.open_dataset('Rtraj')
         ds = af.open_dataset('Sprof')
+
+        ds = af.open_dataset('Sprof', netCDF4=True)  # Return a netCDF4 Dataset instead of an xarray
 
     .. code-block:: python
         :caption: Other attributes and methods