+++ argopy/stores/implementations/http.py
@@ -15,11 +15,13 @@ import warnings
 import io
 from pathlib import Path
 from functools import lru_cache
+from netCDF4 import Dataset
 
 from ...options import OPTIONS
 from ...errors import InvalidMethod, DataNotFound
 from ...utils import Registry, UriCName
-from ...utils.transform import (
+from ...utils import has_aws_credentials
+from ...utils import (
     drop_variables_not_in_all_datasets,
     fill_variables_not_in_all_datasets,
 )
@@ -219,7 +221,7 @@ class httpstore(ArgoStoreProto):
              Options passed to :func:`httpstore.download_url`
 
         xr_opts: dict, default={}
-             Options passed to :func:`xarray.open_dataset`
+             Options passed to :func:`xarray.open_dataset`. This will be ignored if the ``netCDF4` option is set to True.
 
         Other Parameters
         ----------------
@@ -229,9 +231,12 @@ class httpstore(ArgoStoreProto):
         akoverwrite: bool, optional
             Determine if kerchunk data should be overwritten or not. This is passed to :meth:`ArgoKerchunker.to_kerchunk`.
 
+        netCDF4: bool, optional, default=False
+            Return a :class:`netCDF4.Dataset` object instead of a :class:`xarray.Dataset`
+
         Returns
         -------
-        :class:`xarray.Dataset`
+        :class:`xarray.Dataset` or :class:`netCDF4.Dataset`
 
         Raises
         ------
@@ -296,20 +301,26 @@ class httpstore(ArgoStoreProto):
             from .. import ArgoKerchunker
 
             if "ak" not in kwargs:
-                self.ak = ArgoKerchunker(
-                    store="local", root=Path(OPTIONS["cachedir"]).joinpath("kerchunk")
-                )
+                self.ak = ArgoKerchunker()
+                if self.protocol == 's3':
+                    storage_options = {'anon': not has_aws_credentials()}
+                else:
+                    storage_options = {}
+                self.ak.storage_options = storage_options
             else:
                 self.ak = kwargs["ak"]
 
-            if self.ak.supported(url):
+            if self.ak.supported(url, fs=self):
                 xr_opts = {
                     "engine": "zarr",
                     "backend_kwargs": {
                         "consolidated": False,
                         "storage_options": {
-                            "fo": self.ak.to_kerchunk(url, overwrite=akoverwrite),
+                            "fo": self.ak.to_reference(url,
+                                                       overwrite=akoverwrite,
+                                                       fs=self),  # codespell:ignore
                             "remote_protocol": fsspec.core.split_protocol(url)[0],
+                            "remote_options": self.ak.storage_options
                         },
                     },
                 }
@@ -318,10 +329,18 @@ class httpstore(ArgoStoreProto):
                 warnings.warn(
                     "This url does not support byte range requests so we cannot load it lazily, falling back on loading in memory."
                 )
-                log.debug("This url does not support byte range requests: %s" % url)
+                log.debug("This url does not support byte range requests: %s" % self.full_path(url))
                 return load_in_memory(
                     url, errors=errors, dwn_opts=dwn_opts, xr_opts=xr_opts
                 )
+
+        netCDF4 = kwargs.get("netCDF4", False)
+        if lazy and netCDF4:
+            if errors == "raise":
+                raise ValueError("Cannot return a netCDF4.Dataset object in lazy mode")
+            elif errors == "ignore":
+                log.error("Cannot return a netCDF4.Dataset object in lazy mode")
+                return None
 
         if not lazy:
             target, _ = load_in_memory(
@@ -337,11 +356,16 @@ class httpstore(ArgoStoreProto):
             )
 
         if target is not None:
-            ds = xr.open_dataset(target, **xr_opts)
+            if not netCDF4:
+                ds = xr.open_dataset(target, **xr_opts)
 
-            if "source" not in ds.encoding:
-                if isinstance(url, str):
-                    ds.encoding["source"] = url
+                if "source" not in ds.encoding:
+                    if isinstance(url, str):
+                        ds.encoding["source"] = self.full_path(url)
+
+            else:
+                target = target if isinstance(target, bytes) else target.getbuffer()
+                ds = Dataset(None, memory=target, diskless=True, mode='r')
 
             self.register(url)
             return ds
@@ -656,7 +680,7 @@ class httpstore(ArgoStoreProto):
 
         if "lazy" in open_dataset_opts and open_dataset_opts["lazy"] and concat:
             warnings.warn(
-                "Lazy openning and concatenate multiple netcdf files is not yet supported. Ignoring the 'lazy' option."
+                "Lazy opening and concatenate multiple netcdf files is not yet supported. Ignoring the 'lazy' option."
             )
             open_dataset_opts["lazy"] = False