+++ argopy/utils/__init__.py
@@ -19,7 +19,7 @@ from .checkers import (  # noqa: F401
     has_aws_credentials,
 )
 from .casting import DATA_TYPES, cast_Argo_variable_type, to_list
-from .decorators import deprecated, doc_inherit
+from .decorators import deprecated, doc_inherit, register_accessor
 from .lists import (
     list_available_data_src,
     list_available_index_src,
@@ -31,6 +31,7 @@ from .lists import (
     list_radiometry_variables,
     list_radiometry_parameters,
     list_gdac_servers,
+    shortcut2gdac,
 )
 from .caching import clear_cache, lscache
 from .monitored_threadpool import MyThreadPoolExecutor as MonitoredThreadPoolExecutor
@@ -62,6 +63,7 @@ from .transform import (
 from .format import argo_split_path, format_oneline, UriCName, redact
 from .loggers import warnUnless, log_argopy_callerstack
 from .carbon import GreenCoding, Github
+from . import optical_modeling
 
 import importlib
 path2assets = importlib.util.find_spec('argopy.static.assets').submodule_search_locations[0]
@@ -93,6 +95,7 @@ __all__ = (
     # Decorators:
     "deprecated",
     "doc_inherit",
+    "register_accessor",
     # Lists:
     "list_available_data_src",
     "list_available_index_src",
@@ -104,6 +107,7 @@ __all__ = (
     "list_radiometry_variables",
     "list_radiometry_parameters",
     "list_gdac_servers",
+    "shortcut2gdac",
     # Cache management:
     "clear_cache",
     "lscache",
@@ -145,4 +149,6 @@ __all__ = (
     # Carbon
     "GreenCoding",
     "Github",
+    # Optical modeling
+    "optical_modeling",
 )